requirements.txt
Flask
numpy
opencv-python
RPi.GPIO
ultralytics
supervision
picamera2
